refactor: pull odata into its own resource file.
* also mark the json data output as json.
