+++ lib/service.js
@@ -29,6 +29,7 @@ module.exports = (container) => {
   // RESOURCES
 
   require('./resources/field-keys')(service, container);
+  require('./resources/odata')(service, container);
   require('./resources/forms')(service, container);
   require('./resources/users')(service, container);
   require('./resources/sessions')(service, container);
