+++ lib/util/http.js
@@ -298,6 +298,7 @@ const contentType = (type) => (result) => (_, response) => {
 };
 const xml = contentType('application/xml');
 const atom = contentType('application/atom+xml');
+const json = contentType('application/json');
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -311,6 +312,6 @@ module.exports = {
   sendError,
   versionParser, sessionParser, fieldKeyParser, headerOptionsParser,
   getOrElse, getOrReject, getOrNotFound, rejectIf,
-  success, contentType, xml, atom
+  success, contentType, xml, atom, json
 };
